requirements.txt
boto==2.42.0
Django==1.10
django-social-auth==0.7.28
django-storages==1.5.0
djangorestframework==3.5.3
httplib2==0.9.2
Markdown==2.6.7
oauth2==1.9.0.post1
oauthlib==2.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
